Initialize cluster without emitting started event
This is the first in a series of commits to refactor the MongoDB
charm.

This commit replaces the mongodb_started event with a direct call
to its event handler, since both the source of the event and the
event handler are in the same class, and there was no clear benefit
of an asynchoronous call.