Removed rasing exceptions from event handlers
Event handlers are not expected to raise exceptions. In this particular
case the exception was being raised when an event was received by a unit
that was not the leader unit. This is entirely normal and did not
warrant and exception. Hence the exception was replaced but a normal
return.