Remove custom events
This commit removes events custom events emmitted by the charm that are
handled by the charm itself. In order to remove such events most of the
significant functionality of the MongoDBCluster class was mingrated into
the MongoDBCharm class. In subsequent commits the duplicated code will
be removed from the MongoDBCluster class and the custom events will
be redefined so that they are meant to be handled by Charms related to
this charm (as opposed to the charm itself).