Refined MongoDB Cluster class
This commit refines the purpose of the MongoDBCluster class. This class
is meant to be a utility to raise events that are handled by related
charms. The events emmited at present are the same as they were in the
unrefactored charm. In a subsequent commit an instance of this class
will be created in the MongoDBCharm class which will raise the custom
events as required.