Renamed ready state to cluster initialized
This commit renames the ready state variable to cluster_initialized.
The purpose of this renaming was to reduce the number of variables
and methods denoted as "ready". Moreover the ready state variable
was tracking if the MongoDB cluster had already been initialized
(using the "replSetInitiate" command). This renaming made the
intent of this state variable more transparent and alliviated any
confusion with other "ready" indicator variables and methods.