Remove event based busy loop
This commit removes an event based busy loop that was being used by the
charm to initialize the mongodb replica set on startup. The event
loop was implemented by setting application data and then deferring
the start event. Even though the application data that was set has not
changed, the deffered start event was called soon after due to config
changed events triggered as a result of setting the application data. In
fact this behaviour was a bug in Juju, since setting the same
application data should not result in relation changed events being
triggered. A future release of Juju is expected to fix this bug, and
hence the workflow based on it, used by this charm would then stop
working.

As a concequence of removing this logic from the cycle time of
repeatedly trying to initialize mongodb replicaset as part of the start
event handler now happens much slower on the scale of the update status
interval. However this slowness is only visible if a single mongodb unit
is deployed initially. If multiple mongodb units are deployed then since
many more events are generated, mongodb initialization happens fairly
quickly.