Removed redundant replica set initialized check
As part of continuing refactoring this commit removes the replica set
initialized property that was previously used to check if mongodb has
been initialized. This function is now replaced by the state variable
mongodb_initialized. The value of this state variable is set only once
in the start event handler when the call to mongodb initialize API
succeeds. The state variable never needs to be changed after this.