Removed peer relation member variable
This commit removes the member variable that stores the operator
framework peer relation object. The purpose of this change is to
facilitate writing of unit tests. Even though every time an event
handler is invoked in a charm, the Charm class is constructed and
initialized each time, the same behaviour is not implemented in the
Operator Frameworks testing harness. Hence the state of member
variables may become state when writing unit tests.