Added MongoProvider and MongoConsumer
This commit adds MongoProvider and MongoConsumer objects that
are derived from ops.relation.Provider and ops.relation.Consumer. These
two objects manage the relationship between the MongoDB charm and any
charm using the services of the MongoDB charm. The MongoConsumer object
is part of a charm library provided by the MongoDB charm.