Enable user and replica authentication
The commit implements authentication for users and replica sets.
Users need a password to access MongoDB. Replicas need a security
key to be included into a replica set. By default only a single
user "root" is created. The password for this user is autogenerated
and stored in application relation data. The password may be viewed
by the administrator using `juju show-unit mongodb/leader`.

Replica keyfiles are also autogenerated and mounted into each
container.