Removed obsolete bundle
This commit removes the bundle that is now obsolete given the
many changes in this charm since it was added. In the future
a new bundle may be provided. Perhaps one use case for a
bundle is as a means to support sharded clusters. Sharding
may be supported by making the mongodb charm deployment
configurable to launch either the mongod or the mongos
daemons. Then the bundle may launch some mongod and some mongos
instances.