Remove stored state from MongoProvider
This commit removes the use of stored state from the MongoProvider
object and replaces it with peer application relation data. This
is because stored state is unit local and hence application state
must be persisted using relation data only. MonogProvider needs
to persist the list of consumer charm relation ids and associated
database user credentials.