Restart Mongodb when service is not ready
It has been observed that sometimes the MongoDB workload
fails to start, hence the charm is stuck in a mongodb "service not ready
yet" waiting status. This commit adds a temporary fix by restarting
the workload container if this state is observed. In the future
this commit can be reverted when Pebble automatically restarts
containers that have not started.