Sets the db relation data on start (#29)
* Fixes the test_mongoprovider tests

A previous change removed the dependency on ops.relation, but
the unit tests were not updated, which resulted in them failing.

* Adds a unit test for checking the database relations on start

Currently, the MongoDB charm does not set any relation data into the
related charms if they were related before the charm finished starting.

This unit test can be used to verify this aspect, and it should be updated
when this issue is addressed.

* Sets the db relation data on start

If another charm was related to this MongoDB charm before it
finished starting, then the MongoDB would not have set the
needed relation data. In order to prevent this, the relations
had to be created after the MongoDB became active.

This should address the issue. Added a unit test to verify this
behaviour.