mongoprovider: Use the relation IDs as strings instead of ints
"consumers" is a dictionary in which we're storing credentials
based on the Relation ID, and we're storing it as a json dump,
and we're looking into to it whenever we're trying to determine
if we have a new consumer.

However, json.dumps converts integers into strings, which means
that we can't rely on Relation IDs being integers.

This convers the Relation IDs used in mongoprovider.py into strings.