Trap operation failure on shutdown
This commit ensures that the pymongo OperationFailure
exception does not percolate upto Juju and result in an
error message in the Juju status. At present the commit
only logs a warning and ignores the exception. A future
commit will address the issue of how to deal with
database shutdown failure more effectively.