Workarounds for breaking changes in PyMongo and MongoDB (#34)
* Trap operation failure on shutdown

This commit ensures that the pymongo OperationFailure
exception does not percolate upto Juju and result in an
error message in the Juju status. At present the commit
only logs a warning and ignores the exception. A future
commit will address the issue of how to deal with
database shutdown failure more effectively.

* Update replica set hosts on reconfiguration

This commit ensures that the list of current replica set hosts
is kept up to date in the peer relation data bag.

* Pin requirements due to broken backward compatibility

This commit pins the PyMongo requirements. This is necessary
because PyMongo 4 and MongoDB 5 have breaking changes in the
way replica sets are initialized. A future commit will update
this charms codebase and remove this pinning.