Moved key file from persistent to container storage
The MongoDB charm was initaly built as a pod spec charm. The
mongod daemon expects its key file to be present before it is
started. Hence the pod spec charm used a Kubernetes init container
and a persistent storage to create and populate the key file.
This is no longer necessary for a side car charm, since key files
can be pushed to the workload container before starting the
workload. This is exactly the workflow implemented by this
commit.

Closes: #16