Trap additional exception when checking ready status (#40)
The MongoDB charm checks the status of MongoDB prior to
attempting replica set initialization. This check uses
a try and fail approach. The failure is detected through
exceptions rasied which are trapped. Previous version of
PyMongo raised ServerSelectionTimeout exception. The
current version raises OperationFailure exception. This
commit updates the code to trap the new exception.

Closes: #36