WIP(upgrades): Code for pre upgrade checks.
This code is committed and left as is for now.
This has been developed in order to ensure that pre upgrade checks are
working as expected.
It uses some reusable helpers from MongoDB VM for the upgrader
(MongoDBUpgrade) and for the upgrade (KubernetesUpgrade).
Since we need to read the stateful set, it's necessary to trust the
applications we are deploying, hence the need to
`juju deploy --trust`.
There's an error catcher for that which will give instructions in the
status.
As of my tests, this works on replicaset and sharded cluster.